small refactor to warmup for the day
